Update grunt config to enable build for both crx and wgt
Also updated package task so it's a multitask, to enable
multiple package types (e.g. in this case .wgt)
to be built from a single grunt config.

For the crx build, grunt crx will copy all of the required
files to build/crx. However, I don't currently have a grunt
task to produce valid crx packages, so you have to load
the build/crx folder as a Chrome developer extension to see it
working.